+++ InterfaceProperties/Interface/IRegularPolygon.cs
@@ -7,9 +7,9 @@
 
         // Note: These two lines of code are equivalent:
         //       a calculated property with only a getter.
-        //int Perimeter { get => NumberOfSides * SideLength; }
-        int Perimeter => NumberOfSides * SideLength;
+        //double Perimeter { get => NumberOfSides * SideLength; }
+        double Perimeter => NumberOfSides * SideLength;
 
-        int Area { get; }
+        double Area { get; }
     }
 }
